Separated IndexPath extension.
